Rarity combo box properly displays possible values.
When changing the selected item in the drop down, the ID for the rarity in the
familiar type is change.  Presumably, Entity Framework will properly update
the related Rarity object when this is saved.

Also added New, Save, and Delete buttons.  Only the Save button is hooked up
to anything nontrivial, and its event handler is just for testing purposes.